// CPL/auth/apps/grpc-internal/src/channel/internal-channel.enum.ts
export enum CHANNEL_MESSAGE {
  ADD_FAIL = 'CHANNEL.ADD_FAIL',
  NOT_FOUND = 'CHANNEL.NOT_FOUND',
  NOT_EXISTS_SOME_TAGS = 'CHANNEL.NOT_EXISTS_SOME_TAGS',
  EXISTS = 'CHANNEL.EXISTS',
  UPDATE_FAIL = 'CHANNEL.UPDATE_FAIL',
  DELETE_SUCCESS = 'CHANNEL.DELETE_SUCCESS',
  DELETE_FAIL = 'CHANNEL.DELETE_FAIL',
}